requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
joblib
python-multipart
pydantic>=2,<3
pyyaml
pytest
